Removing the linter
